Feature - Movement stock - Create trace of movement of stock in each action

+++ src/main/java/com/inventory_system/backend/model/StockMovement.java
@@ -20,7 +20,7 @@ public class StockMovement {
     @Enumerated(EnumType.STRING)
     private MovementType movementType;
     @Column
-    private Integer cant;
+    private Long cant;
     /*Relations*/
     @ManyToOne(fetch = FetchType.EAGER)
     private User user;
